feat: 매칭 확인 페이지 및 API 구현
- 새로운 ConfirmationView 컴포넌트를 추가하여 매칭 확인 UI를 구성
- 매칭 데이터 API를 추가하여 예약 ID에 따른 매칭 정보를 반환
- 예약 확인 페이지에서 매칭 데이터 로딩 및 에러 처리 로직 구현
- BottomSection 컴포넌트에서 매니저 선택 후 예약 생성 로직 개선
