+++ apps/src/widgets/manager/BottomSection.tsx
@@ -1,8 +1,7 @@
 'use client';
 
 import { useRouter } from 'next/navigation';
-import { useEffect, useState } from 'react';
-import { MAX_MANAGER_COUNT } from '@/widgets/manager/model/manager';
+import { useState } from 'react';
 import { MANAGER_NAMES } from '@/widgets/manager/model/manager';
 
 interface BottomSectionProps {
@@ -12,21 +11,11 @@ interface BottomSectionProps {
 
 export function BottomSection({ selectedManagers, reservationId }: BottomSectionProps) {
   const router = useRouter();
-  const [currentReservationId, setCurrentReservationId] = useState<string | null>(null);
-
-  useEffect(() => {
-    // 처음 예약의 경우 localStorage에서 reservationId 가져오기
-    if (!reservationId && typeof window !== 'undefined') {
-      const storedId = localStorage.getItem('currentReservationId');
-      setCurrentReservationId(storedId);
-    } else {
-      setCurrentReservationId(reservationId || null);
-    }
-  }, [reservationId]);
+  const [isLoading, setIsLoading] = useState(false);
 
   const getSelectedManagerNames = () => {
     if (selectedManagers.length === 0) return '';
-    const names = selectedManagers.map(id => MANAGER_NAMES[id as keyof typeof MANAGER_NAMES]);
+    const names = selectedManagers.map(id => MANAGER_NAMES[id as keyof typeof MANAGER_NAMES] || '알 수 없는 매니저');
     return `${names.join(', ')} 매니저가 선택되었습니다`;
   };
 
@@ -35,37 +24,52 @@ export function BottomSection({ selectedManagers, reservationId }: BottomSection
       alert('최소 1명의 매니저를 선택해주세요.');
       return;
     }
+    
+    setIsLoading(true);
 
     try {
-      // 매니저 선택 완료 후 새 예약 생성
-      // const response = await fetch('/api/reservations', {
-      //   method: 'POST',
-      //   headers: { 'Content-Type': 'application/json' },
-      //   body: JSON.stringify({ 
-      //     selectedManagers,
-      //     reservationId: currentReservationId // localStorage에서 가져온 ID 또는 prop으로 받은 ID
-      //   })
-      // });
-      // const { reservationId: newReservationId } = await response.json();
-      
-      console.log('Creating reservation with managers:', selectedManagers);
-      if (currentReservationId) {
-        console.log('Using reservation ID:', currentReservationId);
+      let finalPayload;
+
+      if (reservationId) {
+        finalPayload = {
+          selectedManagers,
+          reservationId,
+        };
+      } else {
+        const pendingReservationJSON = typeof window !== 'undefined' ? localStorage.getItem('pendingReservation') : null;
+        if (!pendingReservationJSON) {
+          throw new Error('예약 정보를 찾을 수 없습니다. 예약 과정을 다시 시작해 주세요.');
+        }
+        const reservationDetails = JSON.parse(pendingReservationJSON);
+        finalPayload = {
+          ...reservationDetails,
+          selectedManagers,
+        };
+      }
+
+      const response = await fetch('/api/reservations', {
+        method: 'POST',
+        headers: { 'Content-Type': 'application/json' },
+        body: JSON.stringify(finalPayload)
+      });
+
+      if (!response.ok) {
+        const errorData = await response.json().catch(() => ({}));
+        throw new Error(errorData.message || '예약 및 매칭 생성에 실패했습니다.');
       }
       
-      // 실제 예약 ID 사용 (localStorage에서 가져온 것 또는 prop으로 받은 것)
-      const finalReservationId = currentReservationId || 'CL-20230510-1234';
+      const { reservationId: newReservationId } = await response.json();
       
-      // localStorage 정리 (처음 예약 완료 후)
       if (!reservationId && typeof window !== 'undefined') {
-        localStorage.removeItem('currentReservationId');
+        localStorage.removeItem('pendingReservation');
       }
       
-      // 예약 상세 페이지로 이동
-      router.push(`/reservation/${finalReservationId}/confirmation`);
+      router.push(`/reservation/${newReservationId}/confirmation`);
     } catch (error) {
-      console.error('매니저 선택 제출 오류:', error);
-      alert('매니저 선택 중 오류가 발생했습니다.');
+      console.error('매니저 선택 및 예약 생성 제출 오류:', error);
+      alert(error instanceof Error ? error.message : '오류가 발생했습니다.');
+    } finally {
+      setIsLoading(false);
     }
   };
 
@@ -84,11 +88,11 @@ export function BottomSection({ selectedManagers, reservationId }: BottomSection
           <div className="py-4">
             <button
               onClick={handleNext}
-              disabled={selectedManagers.length === 0}
+              disabled={selectedManagers.length === 0 || isLoading}
               className="w-full h-14 rounded-2xl font-semibold text-white text-lg bg-primary disabled:bg-slate-200 disabled:text-slate-400 transition-colors"
             >
-              다음
-              {selectedManagers.length > 0 && (
+              {isLoading ? '처리 중...' : '다음'}
+              {selectedManagers.length > 0 && !isLoading && (
                 <span className="ml-1">›</span>
               )}
             </button>
